services/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
pydantic-settings
python-dotenv
redis
bcrypt
PyJWT
pytest
httpx
